feat: move offer expiration to visibility options
